Up encoding limits(spoiler: doesnt work) and cap request limit to 2^60

+++ src/main/java/appeng/client/gui/me/items/SetProcessingPatternAmountScreen.java
@@ -99,6 +99,6 @@ public class SetProcessingPatternAmountScreen<C extends PatternEncodingTermMenu>
     }
 
     private long getMaxAmount() {
-        return Integer.MAX_VALUE;
+        return Long.MAX_VALUE;
     }
 }
